+++ src/webpack/plugin/applyDefaults.js
@@ -52,6 +52,7 @@ const applyDefaults = ({
   distributionFolder = 'lib',
   enableDebugLogging = false,
   additionalConfig = {},
+  validateUrls = process.env.validateUrls || false,
 } = {}) => ({
   primaryNavigationItems,
   additionalSearchDirectories,
@@ -69,6 +70,7 @@ const applyDefaults = ({
   enableDebugLogging,
   namespace: npmPackage.name,
   additionalConfig,
+  validateUrls,
 });
 
 module.exports = applyDefaults;