+++ src/webpack/plugin/siteLoaderUtils.js
@@ -1,5 +1,10 @@
 const rehypeSlug = require('rehype-slug');
 const rehypeUrl = require('rehype-urls');
+const rehypeUrlInspector = require('@jsdevtools/rehype-url-inspector');
+const { XMLHttpRequest } = require('xmlhttprequest');
+const logging = require('webpack/lib/logging/runtime');
+
+const logger = logging.getLogger('terra-dev-site');
 
 const babelLoader = {
   loader: 'babel-loader',
@@ -8,7 +13,7 @@ const babelLoader = {
   },
 };
 
-const getMdxLoader = (publicPath) => ({
+const getMdxLoader = (publicPath, validateUrls) => ({
   loader: '@mdx-js/loader',
   options: {
     rehypePlugins: [
@@ -21,6 +26,20 @@ const getMdxLoader = (publicPath) => ({
           return `${publicPath}${url.pathname.slice(1)}`;
         }
         return url.href;
+      }],
+      [rehypeUrlInspector, {
+        inspectEach: (url) => {
+          if (validateUrls === 'true' && (url.url.includes('https') || url.url.includes('http'))) {
+            const xmlHttp = new XMLHttpRequest();
+            xmlHttp.open('GET', url.url, false);
+            xmlHttp.onloadend = () => {
+              if (xmlHttp.status === 404 && new RegExp(/Not Found/gi).test(xmlHttp.responseText)) {
+                logger.warn('Warning! Broken Link', url.url, 'in', url.file.history[0], 'at line:', url.node.position.start.line);
+              }
+            };
+            xmlHttp.send(null);
+          }
+        },
       }],
     ],
   },