+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Added `rehype-url-inspector` plugin to warn about broken links.
+
 * Added
   * Added new file `browserslistrc` for Browserslist.
   * Added support for webpack 5.